pyproject.toml
[tool.poetry]
name = "ad-placement-service"
version = "0.1.0"
description = ""
authors = ["Ainur Rakhmanov <ainur-205@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.110.0"
uvicorn = {extras = ["standart"], version = "^0.27.1"}
sqlalchemy = "^2.0.28"
asyncpg = "^0.29.0"
python-dotenv = "^1.0.1"
alembic = "^1.13.1"
passlib = "^1.7.4"
pyjwt = "^2.8.0"


[tool.poetry.group.dev.dependencies]
black = "^24.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
